Final Version

+++ TcpServer/database.cpp
@@ -56,8 +56,16 @@ QString DataBase::Register(const char *name, const char *pwd)
 {
     QString sql = QString("insert into account(username,password) values(\'%1\',\'%2\')").arg(name).arg(pwd);
     QSqlQuery query;
-    if (query.exec(sql))
+    qDebug()<<"SQL: "<<sql;
+
+    if (query.exec(sql))// If register success, then initial the root folder in the folder table !
+    {
+        QString sql = QString("insert into folder(user,name,parent) values(\'%1\','root','-1')").arg(name);
+        QSqlQuery query;
+        qDebug()<<"SQL: "<<sql;
+
         return REGISTER_SUCCESS;
+    }
     else
         return REGISTER_FAIL;
 }
@@ -111,8 +119,22 @@ QString DataBase::Logout(const char *name, const char *pwd)
 }
 
 
-// Folder task1: Query Folder
-// Query the record
+// Folder Task: Get the root ID
+QString DataBase::SearchRootID(const char *name)
+{
+    QString sql = QString("SELECT ID FROM folder WHERE user = '%1' AND name = 'root'").arg(name);
+    qDebug()<<"SQL: "<<sql;
+
+    QSqlQuery query;
+    query.exec(sql);
+    query.next();
+    QString id = query.value("ID").toString() ;
+
+    return id;
+}
+
+
+// Folder task: Query Folder
 QStringList DataBase::RefreshFolder(const char *userName, const char *parent)
 {
     QStringList strList;
@@ -131,8 +153,25 @@ QStringList DataBase::RefreshFolder(const char *userName, const char *parent)
     return strList;
 }
 
-// Folder task2: ADD Folder / File
-// Insert the record
+// Folder task: Refresh file
+QStringList DataBase::RefreshFile(const char *userName, const char *parent)
+{
+    QStringList strList;
+    QString sql = QString("SELECT name FROM file WHERE user = '%1' AND parent = '%2'").arg(userName).arg(parent);
+
+    QSqlQuery query;
+    query.exec(sql);
+
+    while (query.next())
+    {
+        QString unit = query.value("name").toString();
+        strList.append(unit);
+    }
+
+    return strList;
+}
+
+// Folder task: ADD Folder / File
 QString DataBase::AddFolder(const char *userName, const char *parent, const char *folderName)
 {
     QString sql = QString("INSERT INTO folder (user, parent, name) VALUES (\'%1\', \'%2\', \'%3\')").arg(userName).arg(parent).arg(folderName);
@@ -154,21 +193,43 @@ QString DataBase::AddFile(const char *userName, const char *parent, const char *
     qDebug() << "SQL: "<<sql;
 
     if(query.exec(sql))
-        return ADD_FILE_SUCCESS;
+        return UPLOAD_SUCCESS;
     else
-        return ADD_FILE_FAIL1;
+        return UPLOAD_FAIL1;
 
 }
 
 
-// Folder task3: DELETE Folder
-// Delete the record
+// Folder task: DELETE Folder/File
 QString DataBase::DeleteFolder(const char *userName, const char *parent, QStringList idList)
 {
+    if(idList.empty())
+        return DELETE_FOLDER_SUCCESS;
 
     QString sql;
     for (QString id : idList)
+    {
         sql += QString("DELETE FROM folder WHERE user = \'%1\' AND parent = \'%2\' AND ID = \'%3\' ;").arg(userName).arg(parent).arg(id);
+        sql += QString("DELETE FROM folder WHERE user = \'%1\' AND parent = \'%2\' AND ID >= 0 ;").arg(userName).arg(id);
+    }
+    QSqlQuery query;
+    qDebug() << "SQL: "<<sql;
+
+    if(query.exec(sql))
+        return DELETE_FOLDER_SUCCESS;
+    else
+        return DELETE_FOLDER_FAIL;
+
+}
+
+QString DataBase::DeleteFile(const char *userName, const char *parent, QStringList nameList)
+{
+    if(nameList.empty())
+        return DELETE_FOLDER_SUCCESS;
+
+    QString sql;
+    for (QString name : nameList)
+        sql += QString("DELETE FROM file WHERE user = \'%1\' AND parent = \'%2\' AND name = \'%3\' ;").arg(userName).arg(parent).arg(name);
 
     QSqlQuery query;
     qDebug() << "SQL: "<<sql;
@@ -180,8 +241,7 @@ QString DataBase::DeleteFolder(const char *userName, const char *parent, QString
 
 }
 
-// Folder task4: UP Folder
-// SELECT the record
+// Folder task: UP Folder
 QString DataBase::UPFolder(const char *userName, const char *ID)
 {
 
@@ -193,3 +253,36 @@ QString DataBase::UPFolder(const char *userName, const char *ID)
 
     return query.value("parent").toString();
 }
+
+//Folder task: Search Files
+QStringList DataBase::Search(const char *userName, const char *type)
+{
+
+    QStringList strList;
+    QString sql;
+    if (strcmp(type,"Time") == 0)
+        sql = QString("SELECT name FROM file WHERE user = '%1' ORDER BY 'time'  DESC").arg(userName);
+    else if (strcmp(type,"Photo") == 0)
+        sql = QString("SELECT name FROM file WHERE user = '%1' AND  format IN ('jpg', 'png', 'bmp')").arg(userName);
+    else if (strcmp(type,"Music") == 0)
+        sql = QString("SELECT name FROM file WHERE user = '%1' AND format = 'mp3'").arg(userName);
+    else if (strcmp(type,"Video") == 0)
+        sql = QString("SELECT name FROM file WHERE user = '%1' AND format = 'mp4'").arg(userName);
+    else if (strcmp(type,"Document") == 0)
+        sql = QString("SELECT name FROM file WHERE user = '%1' AND format IN ( 'txt', 'ppt', 'pdf', 'doc', 'docx')").arg(userName);
+    else if (strcmp(type,"Else") == 0)
+        sql = QString("SELECT name FROM file WHERE user = '%1' AND format NOT IN ('txt', 'ppt', 'pdf', 'doc', 'docx', 'jpg', 'png', 'mp3', 'mp4')").arg(userName);
+
+    QSqlQuery query;
+    query.exec(sql);
+    qDebug()<<"SQL: "<<sql;
+
+    while (query.next())
+    {
+        QString unit = query.value("name").toString();
+        strList.append(unit);
+    }
+
+    return strList;
+}
+
